Fixed security issue.

+++ yowsup_ext/beholder/layer.py
@@ -30,13 +30,9 @@ class BeholderLayer(YowInterfaceLayer):
 
     #Executes a command if the user is in the allowed_user list. 
     def executeCommand(self, messageProtocolEntity, command):	
-        if messageProtocolEntity.getFrom() in self.__allowed_users:	    
-            status = subprocess.check_output(command)
-            print("Status: "+status)
-            messageProtocolEntity.setBody(status)
-        else:
-            print("Not allowed user '%s'" % messageProtocolEntity.getFrom())
-            messageProtocolEntity.setBody("Authorization check failed!")
+        status = subprocess.check_output(command)
+        print("Status: "+status)
+        messageProtocolEntity.setBody(status)
         self.toLower(messageProtocolEntity.forward(messageProtocolEntity.getFrom()))
 
     @ProtocolEntityCallback("message")
@@ -110,35 +106,41 @@ class BeholderLayer(YowInterfaceLayer):
         self.toLower(entity.ack())
 
     def onTextMessage(self,messageProtocolEntity):
-        if "help" in string.lower(messageProtocolEntity.getBody()):
-            self.showHelp(messageProtocolEntity)
-        #Node list
-        elif "hello" in string.lower(messageProtocolEntity.getBody()): 
-            self.sayHello(messageProtocolEntity)
-        #Enable/Disable node if needed.
-        elif "unselect" in string.lower(messageProtocolEntity.getBody()):
-            self.unselectNode(messageProtocolEntity)
-        elif "select" in string.lower(messageProtocolEntity.getBody()):
-            if string.lower(self.__alias) in string.lower(messageProtocolEntity.getBody()):
-                self.selectNode(messageProtocolEntity)
-        elif "disable" in string.lower(messageProtocolEntity.getBody()):
-            if self.__node_selected == True:
-                self.disableNode(messageProtocolEntity)
-        elif "enable" in string.lower(messageProtocolEntity.getBody()):
-            if self.__node_selected == True:
-                if self.__node_enabled == False :
-                    self.enableNode(messageProtocolEntity)
-        # Execute command if possible.
-        elif string.lower(messageProtocolEntity.getBody()) == "reboot":
-            if self.__node_selected == True :
-                command = ['sudo', 'reboot']
-                self.executeCommand(messageProtocolEntity, command)
+        if messageProtocolEntity.getFrom() in self.__allowed_users:
+            if "help" in string.lower(messageProtocolEntity.getBody()):
+                self.showHelp(messageProtocolEntity)
+            #Node list
+            elif "hello" in string.lower(messageProtocolEntity.getBody()): 
+                self.sayHello(messageProtocolEntity)
+            #Enable/Disable node if needed.
+            elif "unselect" in string.lower(messageProtocolEntity.getBody()):
+                self.unselectNode(messageProtocolEntity)
+            elif "select" in string.lower(messageProtocolEntity.getBody()):
+                if string.lower(self.__alias) in string.lower(messageProtocolEntity.getBody()):
+                    self.selectNode(messageProtocolEntity)
+            elif "disable" in string.lower(messageProtocolEntity.getBody()):
+                if self.__node_selected == True:
+                    self.disableNode(messageProtocolEntity)
+            elif "enable" in string.lower(messageProtocolEntity.getBody()):
+                if self.__node_selected == True:
+                    if self.__node_enabled == False :
+                        self.enableNode(messageProtocolEntity)
+            # Execute command if possible.
+            elif string.lower(messageProtocolEntity.getBody()) == "reboot":
+                if self.__node_selected == True :
+                    command = ['sudo', 'reboot']
+                    self.executeCommand(messageProtocolEntity, command)
+            else:
+                # just print info
+                if self.__node_selected == True :
+                    print("Invalid command '%s' from %s" % (messageProtocolEntity.getBody(), messageProtocolEntity.getFrom(False)))
+                    messageProtocolEntity.setBody("Invalid command '%s'." % messageProtocolEntity.getBody() )
+                    self.toLower(messageProtocolEntity.forward(messageProtocolEntity.getFrom()))
         else:
-            # just print info
-            if self.__node_selected == True :
-                print("Invalid command '%s' from %s" % (messageProtocolEntity.getBody(), messageProtocolEntity.getFrom(False)))
-                messageProtocolEntity.setBody("Invalid command '%s'." % messageProtocolEntity.getBody() )
-                self.toLower(messageProtocolEntity.forward(messageProtocolEntity.getFrom()))
+            print("Not allowed user '%s'" % messageProtocolEntity.getFrom())
+            messageProtocolEntity.setBody("Authorization check failed!")
+            self.toLower(messageProtocolEntity.forward(messageProtocolEntity.getFrom()))
+
 
     #Media messages cannot send commands. 
     def onMediaMessage(self, messageProtocolEntity):
